requirements.txt
asgiref==3.4.1
Django==4.0
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
importlib-metadata==4.8.2
Markdown==3.3.6
numpy==1.21.4
openpyxl==3.0.9
pandas==1.3.5
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
zipp==3.6.0
